# core/docs/swagger.yaml
definitions:
...
  internal_api_handlers.InterfaceEditRequest:
    properties:
      ip:
        type: string
      management:
        items:
          type: string
        type: array
      state:
        type: string
      zone:
        type: string
    type: object
...
  internal_api_handlers.NatCreateRequest:
    properties:
      description:
        type: string
      dst-addr:
        type: string
      dst-zone:
        type: string
      out-interface:
        type: string
      service:
        type: string
      src-addr:
        type: string
      src-zone:
        type: string
      translated-ip:
        type: string
      translated-port:
        type: string
      type:
        type: string
    type: object
  internal_api_handlers.NatEditRequest:
    properties:
      description:
        type: string
      dst-addr:
        type: string
      dst-zone:
        type: string
      out-interface:
        type: string
      service:
        type: string
      src-addr:
        type: string
      src-zone:
        type: string
      translated-ip:
        type: string
      translated-port:
        type: string
      type:
        type: string
    type: object
  internal_api_handlers.NatMoveRequest:
    properties:
      position:
        type: string
      reference_id:
        type: integer
    type: object
...
  sentinelos_core_internal_model.NATRule:
    properties:
      description:
        type: string
      dstAddress:
        type: string
      dstZone:
        $ref: '#/definitions/sentinelos_core_internal_model.Zone'
      id:
        type: integer
      outInterface:
        type: string
      service:
        type: string
      srcAddress:
        type: string
      srcZone:
        $ref: '#/definitions/sentinelos_core_internal_model.Zone'
      translatedIP:
        type: string
      translatedPort:
        type: string
      type:
        $ref: '#/definitions/sentinelos_core_internal_model.NATType'
    type: object
  sentinelos_core_internal_model.NATType:
    enum:
    - snat
    - dnat-ip
    - dnat-port
    type: string
    x-enum-varnames:
    - TypeSNAT
    - TypeDNATIP
    - TypeDNATPort
...
  sentinelos_core_internal_model.Zone:
    properties:
      color:
        type: string
      interfaces:
        items:
          type: string
        type: array
      name:
        type: string
      networks:
        items:
          type: string
        type: array
      type:
        $ref: '#/definitions/sentinelos_core_internal_model.ZoneType'
    type: object
  sentinelos_core_internal_model.ZoneType:
    enum:
    - l2
    - l3
    type: string
    x-enum-varnames:
    - ZoneL2
    - ZoneL3
  sentinelos_core_internal_system.InterfaceInfo:
    properties:
      ip:
        type: string
      management:
        items:
          type: string
        type: array
      name:
        type: string
      state:
        type: string
      zone:
        type: string
    type: object
  sentinelos_core_pkg_utils.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
  description: Next-Generation Firewall Management API.
  title: SentinelOS API
  version: "1.0"
paths:
...
  /api/interfaces:
    get:
      consumes:
      - application/json
      description: Get a list of all network interfaces.
      parameters:
      - description: Filter by OSI layer (2 or 3)
        in: query
        name: layer
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of interfaces
          schema:
            items:
              $ref: '#/definitions/sentinelos_core_internal_system.InterfaceInfo'
            type: array
        "500":
          description: ERR_SYS_4001 Internal server error
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: List Interfaces
      tags:
      - networking
      - interfaces
  /api/interfaces/{name}:
    put:
      consumes:
      - application/json
      description: Update an existing network interface.
      parameters:
      - description: Interface Name
        in: path
        name: name
        required: true
        type: string
      - description: Interface details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.InterfaceEditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: interface updated in candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid JSON or Zone does not exist
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "404":
          description: ERR_NET_1003 Resource not found
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Edit Interface
      tags:
      - networking
      - interfaces
...
  /api/nat:
    get:
      description: Get a list of all NAT rules.
      parameters:
      - description: Filter by NAT type (e.g. snat, dnat-ip)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of NAT rules
          schema:
            items:
              $ref: '#/definitions/sentinelos_core_internal_model.NATRule'
            type: array
        "500":
          description: ERR_SYS_4001 Internal server error
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: List NAT Rules
      tags:
      - security
      - nat
    post:
      consumes:
      - application/json
      description: Create a new NAT rule.
      parameters:
      - description: NAT details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.NatCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: NAT rule added in candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid JSON or NAT type
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Create NAT Rule
      tags:
      - security
      - nat
  /api/nat/{id}:
    delete:
      description: Delete an existing NAT rule.
      parameters:
      - description: NAT Rule ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 'message: NAT rule deleted in candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "404":
          description: ERR_NET_1003 Resource not found
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Delete NAT Rule
      tags:
      - security
      - nat
    put:
      consumes:
      - application/json
      description: Update an existing NAT rule.
      parameters:
      - description: NAT Rule ID
        in: path
        name: id
        required: true
        type: integer
      - description: NAT details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.NatEditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: NAT rule updated in candidate'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid JSON or ID
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "404":
          description: ERR_NET_1003 Resource not found
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Edit NAT Rule
      tags:
      - security
      - nat
  /api/nat/{id}/move:
    post:
      consumes:
      - application/json
      description: Move a NAT rule relative to another rule (top, bottom, before,
        after).
      parameters:
      - description: NAT Rule ID to move
        in: path
        name: id
        required: true
        type: integer
      - description: Move details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.NatMoveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'message: NAT rule moved successfully'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid JSON or Position
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
        "404":
          description: ERR_NET_1003 NAT Rule not found
          schema:
            $ref: '#/definitions/sentinelos_core_pkg_utils.APIError'
      security:
      - ApiKeyAuth: []
      summary: Move NAT Rule
      tags:
      - security
      - nat
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"